Break out ColorComparisonInput.svelte and ColorPresetsInput.svelte
